Convolutions and Gates
Gates are required for doing composition- you need to be able to mix voices based on time. Since I'm adding a gate function, it also makes sense to change the square wave to generate its signal via gates, thus allowing us to configure the duty cycle. This will also be useful for composition.

Since this means there are loads of square wave applications, but we often want to soften our waves, added averaging and convolution functions so that we can "blur" things.